requirements.txt
aiofile==3.9.0
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.0.1
caio==0.9.17
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mongomock==4.3.0
mongomock-motor==0.0.34
motor==3.6.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
rsa==4.9
sentinels==1.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
